mirroring change in 2gb imx8
