+++ .gitmodules
@@ -14,9 +14,10 @@
 	path = Silicon/ARM/MU_TIANO
 	url = https://github.com/mknutsen/mu_silicon_arm_tiano.git
 	branch = release/201903
+[submodule "Common/MU_OEM_SAMPLE"]
+	path = Common/MU_OEM_SAMPLE
+	url = https://github.com/Microsoft/mu_oem_sample.git
+	branch = release/201903
 [submodule "Silicon/ARM/NXP"]
 	path = Silicon/ARM/NXP
 	url = https://github.com/ms-iot/MU_SILICON_NXP.git
-[submodule "Common/MU_OEM_SAMPLE"]
-	path = Common/MU_OEM_SAMPLE
-	url = https://windowspartners.visualstudio.com/MSCoreUEFI/_git/mu_oem_sample
